Don't silence 2nd exceptions.
When an example raises an error and an after or around hook also raises an error, the 2nd error was being silenced, since `set_exception` was a no-op when the exception was already set.  Silencing errors is bad, so we should print the error instead.

See https://github.com/myronmarston/vcr/issues/167#issuecomment-5814249 for the issue that triggered this change.