Disable coveralls from the cukes.
That last commit improved the coveralls percent but the results
are still a bit odd. In our cukes, the rspec code is only run in a
separate process started by aruba using the rspec command. I don't
think that our cukes provide useful info about the coverage percent
due to the separate processes.