Remove failing scenario.
This fails because ruby processes the `-r` option
before the load path for the local (but not installed)
version of rspec-core has been setup, so it fails.

I don't think the documentation value of this scenario
is high enough to warrant spending a long time finding
a solution, so I'm just removing it.