In the rake task, setup load path for rspec-(core|support).
We want to ensure that the currently running version of rspec
gets loaded when the rake task shells out.  Simply loading
the correct executable doesn't guarantee that since it does
a simple `require 'rspec/core'` which could load a different
version (such as a newer gem installation) or fail (such as
when there is no normal gem installation, such as when using
bundler's --standalone mode).