document that include_context takes a block
This patch documents that `include_context` evaluates it's block
argument.

`include_context` evaluating it's block helps express intent.  This is
useful when overriding `let`s that are defined inside the shared
context.  Since they are not visible within the spec file, the block for
`include_context` can be used to indicate where the overridden `let` is
originally defined.